silence "used code" warnings
