+++ rust/healpix-geo-core/src/index/mod.rs
@@ -4,3 +4,9 @@ mod indexing;
 mod ops;
 mod region;
 mod set;
+
+pub use self::geometry::GeometryQuery;
+pub use self::indexers::{ConcreteSlice, LabelIndexer, PositionalIndexer, Slice};
+pub use self::indexing::{Indexing, LabelIndexing, PositionIndexing};
+pub use self::region::CellRegion;
+pub use self::set::SetOperations;
