+++ rust/healpix-geo-core/src/index/set.rs
@@ -1,4 +1,4 @@
-pub(crate) trait SetOperations {
+pub trait SetOperations {
     fn union(&self, other: &Self) -> Self;
     fn intersection(&self, other: &Self) -> Self;
     fn difference(&self, other: &Self) -> Self;
