+++ rust/healpix-geo-core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod ellipsoid;
+pub mod geometry;
 pub mod index;
 pub mod scalar;
 pub mod vectorized;
